+++ app/src/main/java/io/mainframe/hacs/main/MainActivity.java
@@ -219,13 +219,12 @@ public class MainActivity extends AppCompatActivity implements SshUiHandler.OnSh
                 Manifest.permission.ACCESS_NETWORK_STATE,
                 Manifest.permission.ACCESS_WIFI_STATE,
                 Manifest.permission.INTERNET,
-                Manifest.permission.READ_EXTERNAL_STORAGE,
-                Manifest.permission.WRITE_EXTERNAL_STORAGE,
                 Manifest.permission.WAKE_LOCK,
-                Manifest.permission.READ_PHONE_STATE,
+//                Manifest.permission.READ_PHONE_STATE,
                 Manifest.permission.RECEIVE_BOOT_COMPLETED,
                 // to get the ssid
-                Manifest.permission.ACCESS_COARSE_LOCATION
+                Manifest.permission.ACCESS_COARSE_LOCATION,
+                Manifest.permission.ACCESS_FINE_LOCATION,
         };
         if (!EasyPermissions.hasPermissions(this, perms)) {
             Logger.warn("Requesting permission");